Manage Users search functionality works. Manage Customers search functionality works. Alerts placed on delete buttons for Manage Users and Customers

+++ src/Controller/ManageCustomersController.java
@@ -20,11 +20,14 @@ package Controller;
 import DAO.CustomerDAO;
 import Model.Customer;
 import Utilities.DatabaseConnector;
+import Utilities.Validator;
 import com.mysql.jdbc.Connection;
 import java.io.IOException;
 import java.net.URL;
 import java.sql.SQLException;
+import java.util.Optional;
 import java.util.ResourceBundle;
+import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
@@ -33,6 +36,7 @@ import javafx.fxml.Initializable;
 import javafx.scene.Parent;
 import javafx.scene.Scene;
 import javafx.scene.control.*;
+import javafx.scene.control.Alert.AlertType;
 import javafx.scene.control.cell.PropertyValueFactory;
 import javafx.stage.Stage;
 
@@ -102,24 +106,51 @@ public class ManageCustomersController implements Initializable {
     
     @FXML
     void onActionActiveSelected(ActionEvent event) {
-
+        searchTextField.setText("");
+        refreshData();
     }
 
     @FXML
     void onActionAllSelected(ActionEvent event) {
-
+        searchTextField.setText("");
+        refreshData();
     }
     
     @FXML
     void onActionInactiveSelected(ActionEvent event) {
-
+        searchTextField.setText("");
+        refreshData();
     }
 
     @FXML
     void onActionSearch(ActionEvent event) {
+        ObservableList<Customer> customerSearchResultsList = FXCollections.observableArrayList();
 
+        String searchText = searchTextField.getText();
+
+        if (searchText.isEmpty()) {
+            manageCustomersTableView.setItems(customerDAO.query());
+            return;
+        }
+
+        customerSearchResultsList.clear();
+
+        if (Validator.isSearchStringNumber(searchTextField.getText())) {
+            customerSearchResultsList = customerDAO.lookupCustomer(Integer.parseInt(searchText));
+
+        } else {
+            customerSearchResultsList = customerDAO.lookupCustomer(searchText);
+        }
+
+        if (customerSearchResultsList.isEmpty()) {
+            Alert alert = new Alert(Alert.AlertType.INFORMATION, "A customer was not found.");
+            alert.showAndWait();
+        } else {
+            manageCustomersTableView.setItems(customerSearchResultsList);
+        }
     }
 
+    //Go Back button
     @FXML
     void onActionShowDashboard(ActionEvent event) throws IOException {
         displayScreen("/View/Dashboard.fxml", event);
@@ -134,12 +165,16 @@ public class ManageCustomersController implements Initializable {
 
     @FXML
     void onActionDeleteCustomer(ActionEvent event) {
-        try (Connection conn = dc.createConnection()) {
-            customerDAO.remove(manageCustomersTableView.getSelectionModel().getSelectedItem().getCustomerId());
-            refreshData();
-            conn.close();
-        } catch (SQLException | ClassNotFoundException ex) {
-            System.out.println(ex.getMessage());
+        Alert alert = new Alert(AlertType.CONFIRMATION,"Are you sure you want to delete this customer?");
+        ButtonType result = alert.showAndWait().orElse(ButtonType.CANCEL);
+        if(!ButtonType.CANCEL.equals(result)){
+            try (Connection conn = dc.createConnection()) {
+                customerDAO.remove(manageCustomersTableView.getSelectionModel().getSelectedItem().getCustomerId());
+                refreshData();
+                conn.close();
+            } catch (SQLException | ClassNotFoundException ex) {
+                System.out.println(ex.getMessage());
+            }  
         }
     }
 
@@ -162,8 +197,11 @@ public class ManageCustomersController implements Initializable {
             stage.setScene(new Scene(scene));
             stage.show();
 
-        } catch (IOException | NullPointerException ex) {
+        } catch (IOException ex) {
             System.out.println(ex.getMessage());
+        } catch (NullPointerException ex) {
+            Alert alert = new Alert(Alert.AlertType.ERROR, "You must select a customer to edit.");
+            alert.showAndWait();
         }
 
     }
@@ -172,8 +210,14 @@ public class ManageCustomersController implements Initializable {
         ObservableList<Customer> customers;
 
         try (Connection conn = dc.createConnection()) {
-
-            customers = customerDAO.queryWithAddress();
+            
+            if(allRadioButton.isSelected()){
+                customers = customerDAO.queryWithAddress();
+            } else if(activeRadioButton.isSelected()){
+                customers = customerDAO.queryActiveInactive(true);
+            } else {
+                customers = customerDAO.queryActiveInactive(false);
+            }
 
             conn.close();
